Changed my mind, moved the functions into the "csvTools.py" but Am not sure if it will work first try, adding try and except statements for a little while, If calling the remote function from the library fails, use a local (Temprary) copy.
